+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - (Feature) Integration Service Authentication
 - (Improvement) Better panic handling
 - (Feature) PongV1 Integration Service
+- (Feature) Custom Gateway image
 
 ## [1.2.42](https://github.com/arangodb/kube-arangodb/tree/1.2.42) (2024-07-23)
 - (Maintenance) Go 1.22.4 & Kubernetes 1.29.6 libraries